+++ angular-code-input/src/lib/code-input.module.ts
@@ -1,6 +1,7 @@
-import {NgModule} from '@angular/core';
-import {CodeInputComponent} from './code-input.component';
+import {ModuleWithProviders, NgModule} from '@angular/core';
+import {CodeInputComponent } from './code-input.component';
 import {CommonModule} from '@angular/common';
+import {CodeInputComponentConfig, CodeInputComponentConfigToken} from './code-input.component.config';
 
 @NgModule({
   imports: [
@@ -13,4 +14,13 @@ import {CommonModule} from '@angular/common';
     CodeInputComponent
   ]
 })
-export class CodeInputModule {}
+export class CodeInputModule {
+  static forRoot(config: CodeInputComponentConfig): ModuleWithProviders<CodeInputModule> {
+    return {
+      ngModule: CodeInputModule,
+      providers: [
+        {provide: CodeInputComponentConfigToken, useValue: config }
+      ]
+    };
+  }
+}